Fix possible "Unknown system errorno 7" when running on big projects
This is a band-aid for an issue that can happen when you have massive
chunks in your webpack project. The file list sent to the git-bundle-sha
module for calculation can be huge, with each path being absolute. This
can sometimes hit OS argument length limits for system calls.

The patch is to trim up the paths by converting them to be relative to
the current working directory (where git is assumed to be running.)
Should get rid of a lot of redundant characters.